Moved files into package econtools
