+++ project_tests.sh
@@ -6,6 +6,6 @@ virtualenv venv
 source venv/bin/activate
 pip install -r requirements.txt
 pylint -E *.py
-coverage run -m pytest test_feeder.py
-coverage report -m econ_article_feeder.py
+coverage run -m pytest econtools/test_feeder.py
+coverage report -m econtools/*.py
